# combi-uees-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMBI UEES con Laravel 12",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rutas": {
            "get": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Obtener todas las rutas con sus relaciones",
                "operationId": "07f8f66135da1afc0dbdded354b14e48",
                "responses": {
                    "200": {
                        "description": "Lista de rutas completas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ruta"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Crear una nueva ruta",
                "operationId": "7ae787bdba22d5f0225ce1875bc8f0d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rutaNombre",
                                    "rutaLatitud",
                                    "rutaLongitud"
                                ],
                                "properties": {
                                    "rutaNombre": {
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "Ruta Norte"
                                    },
                                    "rutaLatitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -2.1458
                                    },
                                    "rutaLongitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -79.9675
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ruta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ruta"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/rutas/{ruta}": {
            "put": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Actualizar una ruta existente",
                "operationId": "0135e1f85d528d0037e8d135e8ce00b3",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rutaNombre",
                                    "rutaLatitud",
                                    "rutaLongitud"
                                ],
                                "properties": {
                                    "rutaNombre": {
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "Ruta Norte Actualizada"
                                    },
                                    "rutaLatitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -2.146
                                    },
                                    "rutaLongitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -79.968
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ruta actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ruta"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Eliminar una ruta",
                "operationId": "01a00369de29f2118c46d5ea3e099c59",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ruta eliminada exitosamente"
                    },
                    "404": {
                        "description": "Ruta no encontrada"
                    }
                }
            }
        },
        "/api/rutas/{id}/ubicacion-buses": {
            "get": {
                "tags": [
                    "Rutas"
                ],
                "summary": "Obtener conteo de motoristas por ubicación",
                "operationId": "d8b9838ac0d10847791023c09b92c3b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conteo de motoristas en UEES y en la ruta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "motoristasUEES": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "motoristasRuta": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ruta no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Ruta": {
                "title": "Ruta",
                "description": "Información de una ruta de transporte",
                "required": [
                    "rutaID",
                    "rutaNombre",
                    "rutaLatitud",
                    "rutaLongitud",
                    "rutaBusesDisponibles",
                    "rutaBusesTotales"
                ],
                "properties": {
                    "rutaID": {
                        "type": "integer",
                        "example": 1
                    },
                    "rutaNombre": {
                        "type": "string",
                        "maxLength": 30,
                        "example": "Ruta Norte"
                    },
                    "rutaLatitud": {
                        "type": "number",
                        "format": "float",
                        "example": -2.1458
                    },
                    "rutaLongitud": {
                        "type": "number",
                        "format": "float",
                        "example": -79.9675
                    },
                    "rutaBusesDisponibles": {
                        "type": "integer",
                        "example": 2
                    },
                    "rutaBusesTotales": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-26T15:40:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-26T15:45:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}